refactor: Collect beacon knowledge in BeaconModel
